Expose all vendored mojopkgs to all mojo targets
This allows importing any "standard library" package that is vendored
with the mojo compiler without any manual additions to deps.

The biggest consideration here is how we would otherwise handle the
dependency tree of the vendored dependencies themselves. That isn't
easily statically knowable and organically changes between releases. The
trade off here is that when you want to build them all from source,
we'll likely have to have another customization point to make this work