Add GPU detection logic (#12)
There are a few pieces to GPU detection for supporting Mojo. The final
goal is to have a valid value we can pass as
`--target-accelerator=nvidia:80` to `mojo build`. In order to do this
based on the target platform we have to detect the current GPU with
nvidia-smi or rocm-smi, parse the output, and setup various
config_settings to determine which toolchain should be used. If you use
`--platforms=@mojo_host_toolchain` you get the rest of this logic for
free. The currently known supported GPUs are seeded in the
`mojo.gpu_toolchains` module extension, but new ones can be added in
individual projects as well.